Only delete coverage report if it exists
If there is a compilation error when running the tests, the `rm` line at the bottom of this file still attempts to run and returns an annoying error. This change addresses that.

+++ coverage.sh
@@ -8,5 +8,8 @@ fi
 DAIRYSECRET="do-not-use-secrets-like-this-plz" go test github.com/dairycart/dairycart/api -coverprofile=coverage.out -tags test
 go tool cover -html=coverage.out
 
-# delete the new coverage report so I don't accidentally commit it to the repo somehow
-rm coverage.out
+# delete the new coverage report, if it exists, so I don't accidentally commit it to the repo somehow
+if [ -f coverage.out ]
+then
+    rm coverage.out
+fi
